+++ src/Components/Stock/Distribution/DistributionMgmt.js
@@ -7,6 +7,8 @@ import DistributionAdd from './DistributionAdd';
 import DistributionUpdate from './DistributionUpdate';
 import DistributionSearch from './DistributionSearch';
 import DistributionSearchDate from './DistributionSearchDate';
+import Pagination from '../../Customer/modules/PaginationModule';
+import instance from './../../../api/axios';
 
 import "../../Main/Main.css";
 import "./Distribution.css";
